Remove usecases
