+++ Demo/Demo/Networking/DemoMerchantAPI.swift
@@ -43,7 +43,6 @@ final class DemoMerchantAPI {
         return try parse(from: data)
     }
 
-    
     func createApprovalSessionID(accessToken: String, approvalSessionRequest: ApprovalSessionRequest) async throws -> VaultSessionID {
         guard let url = buildPayPalURL(with: "/v2/vault/payment-tokens") else {
             throw URLResponseError.invalidURL
