--- Demo/Demo/UseCases/GetAccessToken.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-final class GetAccessToken {
-
-    func execute() async -> String? {
-        await DemoMerchantAPI.sharedService.getAccessToken(environment: DemoSettings.environment)
-    }
-}
